psv5/proyecto-ps/programacion/contenedor/requirements.txt
django==4.2.11
mysqlclient==2.1.1
gunicorn==20.1.0
python-dotenv==1.0.0
django-extensions
django-simple-captcha
#django-recaptcha>=4.0.0,<5.0.0
requests>=2.25.1,<3.0.0
docker
flask
